app/requirements.txt
coverage==4.5.2
defusedxml==0.5.0
Django==2.1.3
django-tastypie==0.14.2
lxml==4.2.5
mimeparse==0.1.3
model-mommy==1.6.0
python-dateutil==2.7.5
python-mimeparse==1.6.0
pytz==2018.7
selenium==3.141.0
six==1.11.0
South==1.0.2
urllib3==1.24.1
